Update rssh_client , add MA869 band. add TTNv3 server fix LG02 to be able to receive 255 bytes

+++ haserl-ui/root/etc/init.d/DR_enable_rssh.sh
@@ -3,39 +3,32 @@ START=99
 
 start()
 {	
-	host_id="$(uci -q get rssh.rssh.host_id)"
-	host_addr="$(uci -q get rssh.rssh.host_addr)"
-	host_port="$(uci -q get rssh.rssh.host_port)"
-
-	killall ssh
-	sleep 2
-
-	# Connect to server
-	ssh -y -i /etc/dropbear/id_dropbear -o "ExitOnForwardFailure yes" -f -N -T -R $host_port:localhost:22 -K 60 $host_id@$host_addr > /dev/null
+	manual_connect=$(uci get rssh.rssh.manual_connect)
+	
+	if [ ! -f "/etc/rc.d/S99DR_enable_rssh.sh" ] && [ $manual_connect != "1" ];then
+		exit
+	fi
 
 	# Check for successful connection
 	sleep 2
-	check_connect=$(ps | grep -c -e "ssh -y -i /etc/dropbear/id_dropbear")
-	manual_connect=$(uci get rssh.rssh.manual_connect)
+	check_connect=$(ps | grep "ssh" | grep "localhost:22" -c)
 	
-	# If connect failed, and it is an auto connect, start background task to
-	# wait for server to drop prior connection, then connect.
-	if [ $check_connect != "2" ] && [ $manual_connect == "0" ]; then
-		(sleep 300; \
-		ssh -y -i /etc/dropbear/id_dropbear -o "ExitOnForwardFailure yes" -f -N -T -R $host_port:localhost:22 -K 60 $host_id@$host_addr > /dev/null) &
+	# Connecr to Server
+	if [ $check_connect != "1" ]; then
+		host_id="$(uci -q get rssh.rssh.host_id)"
+		host_addr="$(uci -q get rssh.rssh.host_addr)"
+		rssh_id="$(uci -q get rssh.rssh.rssh_id)"
+		killall rssh_client
+		killall ssh
+		rssh_client -s $host_addr -m $rssh_id -u $host_id > /dev/null &
 	fi
-
-	# Reset manual connect flag
-	uci set rssh.rssh.manual_connect="0"
-	uci commit rssh
 }
 
 stop()
 {	
-echo "Manual disconnect" > /tmp/date.txt
+	echo "Manual disconnect" > /tmp/date.txt
 	uci set rssh.rssh.manual_connect="0"
 	uci commit rssh
+	killall rssh_client
 	killall ssh
 }
-
-
